Initial implementation of phxcat
Utilizes the node repl to allow interfacing with the channel. This
prints all messages that come across a channel and does not require to
listen for a specific message.

`socket` and `channel` are available via the repl context for debugging
those. This allows you to listen and perform specific actions on channel
or socket if you so choose.